# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Kapu$ta app endpoints document",
    "version": "1.0.0",
    "title": "Swagger Kapu$ta",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "oksana_stegko@ukr.net"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "kapusta-final-project-nodejs.herokuapp.com",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/transactions/expenses":{
      "post": {
        "tags": [
          "transaction"
        ],
        "summary": "Add new expenses transaction",
        "description": "This can only be done by the logged in user.",
        "operationId": "addExpenses",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "category": {
                  "type":"string",
                  "$ref": "#/definitions/Category/properties/_id"
                },
                "description": {
                  "type": "string",
                  "example": "подарки"
                },
                "value": {
                  "type": "number",
                  "example": 0
                },
                "date":{
                  "type":"string",
                  "example":"1637507578000"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "201 OK. Transaction added!"
                },
                "trans": {
                  "type": "object",
                  "$ref": "#/definitions/Transaction"
                },
                "currentBalance":{
                  "type":"number",
                  "$ref":"#/definitions/User/properties/balance"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    },
    "/transactions/incomes":{
      "post": {
        "tags": [
          "transaction"
        ],
        "summary": "Add new incomes transaction",
        "description": "This can only be done by the logged in user.",
        "operationId": "addIncomes",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "category": {
                  "type":"string",
                  "$ref": "#/definitions/Category/properties/_id"
                },
                "description": {
                  "type": "string",
                  "example": "откуда прибыль"
                },
                "value": {
                  "type": "number",
                  "example": 0
                },
                "date":{
                  "type":"string",
                  "example":"1637507578000"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "201 OK. Transaction added!"
                },
                "trans": {
                  "type": "object",
                  "$ref": "#/definitions/Transaction"
                },
                "currentBalance":{
                  "type":"number",
                  "$ref":"#/definitions/User/properties/balance"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    },
    "/transactions/{id}":{
      "delete": {
        "tags": [
          "transaction"
        ],
        "summary": "Deleted transaction by id",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteTransactionById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The transaction id"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction successfully deleted"
          },
          "404": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "jwt_token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Transaction": {
      "type": "object",
      "properties": {
        "owner": {
          "type": "object",
          "$ref": "#/definitions/User/properties/_id"
        },
        "category": {
          "type": "object",
          "$ref": "#/definitions/Category/properties/_id"
        },
        "date": {
          "type": "string",
          "format":"date-time"
        },
        "year":{
          "type": "number"
        },
        "month": {
          "type": "number"
        },
        "day": {
          "type": "number"
        },
        "description": {
          "type": "string",
          "example": "text"
        },
        "value": {
          "type": "number",
          "example": 0
        },
        "_id":{
          "type":"string",
          "description":"Auto generated by MongoDB"
        }
      },
      "xml": {
        "name": "Transaction"
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Error message"
        }
      }
    }
  }
}